refactor: remove cache module and update dependencies
- Remove `cache` module and related dependencies
- Update imports and paths in affected files
- Simplify `CacheBuilder` and `Cache` struct methods
- Remove unused error variants and options
