+++ Cargo.toml
@@ -5,7 +5,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-cache = { path = "cache" }
 derive_lazy_playlist_child = { path = "derive_lazy_playlist_child" }
 tokio = { version = "1", features = [
     "rt",
@@ -37,6 +36,10 @@ object_store = { version = "0.11", features = ["gcp", "aws", "azure", "http"] }
 moka = { version = "0.12", features = ["future"] }
 clap = { version = "4.5", features = ["derive"] }
 async-stream = "0.3"
+sha2 = "0.10"
+tempfile = "3.18"
+thiserror = "2"
+glob = "0.3"
 
 [dev-dependencies]
 static_assertions = "1.1"
