--- cache/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "cache"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-sha2 = "0.10"
-chrono = "*"
-tempfile = "3.18"
-anyhow = "*"
-serde_json = "*"
-serde = { version = "*", features = ["derive"] }
-thiserror = "2"
-reqwest = { version = "*", features = ["blocking"] }
-log = "*"
-glob = "0.3"
-async-trait = "*"
-tokio = { version = "*", features = [
-    "rt",
-    "net",
-    "macros",
-    "rt-multi-thread",
-    "fs",
-    "io-util",
-] }
-bytes = "*"
-tokio-stream = "*"
-futures = { version = "*", features = ["thread-pool"] }
-tokio-util = { version = "0.7", features = ["codec", "rt", "io", "net"] }
-object_store = { version = "*", features = ["gcp", "aws", "azure", "http"] }
-async-stream = "*"
