+++ src/cache/cache_struct/cache_builder.rs
@@ -2,8 +2,8 @@ use std::default::Default;
 use std::env;
 use std::path::PathBuf;
 
-use crate::Error;
-use crate::{FileDownloader, LocalDownloader};
+use super::super::{FileDownloader, LocalDownloader};
+use super::Error;
 
 use super::Cache;
 
@@ -30,11 +30,6 @@ impl CacheBuilder {
         }
     }
 
-    /// Construct a new `CacheBuilder` with a `ClientBuilder`.
-    pub fn with_file_downloader(file_downloader: Box<dyn FileDownloader>) -> CacheBuilder {
-        CacheBuilder::new().file_downloader(file_downloader)
-    }
-
     /// Set the cache location. This can be set through the environment
     /// variable `RUST_CACHED_PATH_ROOT`. Otherwise it will default to a subdirectory
     /// named 'cache' of the default system temp directory.
@@ -46,13 +41,6 @@ impl CacheBuilder {
     /// Set the `ClientBuilder`.
     pub fn file_downloader(mut self, file_downloader: Box<dyn FileDownloader>) -> CacheBuilder {
         self.config.file_downloader = file_downloader;
-        self
-    }
-
-    /// Set the default freshness lifetime, in seconds. The default is None, meaning
-    /// the ETAG for an external resource will always be checked for a fresher value.
-    pub fn freshness_lifetime(mut self, freshness_lifetime: u64) -> CacheBuilder {
-        self.config.freshness_lifetime = Some(freshness_lifetime);
         self
     }
 
@@ -74,26 +62,5 @@ impl CacheBuilder {
 impl Default for CacheBuilder {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-/// Options to use with [`Cache::cached_path_with_options`].
-#[derive(Default)]
-pub struct Options {
-    /// An optional subdirectory (relative to the cache root) to cache the resource in.
-    pub subdir: Option<String>,
-}
-
-impl Options {
-    pub fn new(subdir: Option<&str>) -> Self {
-        Self {
-            subdir: subdir.map(String::from),
-        }
-    }
-
-    /// The the cache subdirectory to use.
-    pub fn subdir(mut self, subdir: &str) -> Self {
-        self.subdir = Some(subdir.into());
-        self
     }
 }
