+++ src/cache/error.rs
@@ -7,15 +7,6 @@ pub enum Error {
     #[error("Resource not found at the location ({0})")]
     ResourceNotFound(String),
 
-    /// Arises when the resource looks like a URL, but is invalid.
-    #[error("Unable to parse resource URL ({0})")]
-    InvalidUrl(String),
-
-    /// Arises when the cache is being used in offline mode, but it couldn't locate
-    /// any cached versions of a remote resource.
-    #[error("Offline mode is enabled but no cached versions of resouce exist ({0})")]
-    NoCachedVersions(String),
-
     /// Arises when the cache is corrupted for some reason.
     ///
     /// If this error occurs, it is almost certainly the result of an external process
@@ -24,27 +15,23 @@ pub enum Error {
     #[error("Cache is corrupted ({0})")]
     CacheCorrupted(String),
 
-    /// Arises when a resource is treated as archive, but the extraction process fails.
-    #[error("Extracting archive failed ({0})")]
-    ExtractionError(String),
-
     /// Any IO error that could arise while attempting to cache a remote resource.
     #[error("An IO error occurred")]
-    IoError(#[from] std::io::Error),
+    Io(#[from] std::io::Error),
 
     /// Failed to get the object storage.
     #[error("Failed to get object storage")]
-    ObjectStorageError(#[from] object_store::Error),
+    ObjectStorage(#[from] object_store::Error),
 
     /// Failed to parse a object storage path.
     #[error("Failed to parse object storage path")]
-    ObjectStoragePathError(#[from] object_store::path::Error),
-
-    /// An HTTP error that could occur while attempting to fetch a remote resource.
-    #[error(transparent)]
-    HttpError(#[from] reqwest::Error),
+    ObjectStoragePath(#[from] object_store::path::Error),
 
     /// A method that should never be called was called.
     #[error("Method not implemented")]
     NotImplemented,
 }
+
+// TODO An HTTP error that could occur while attempting to fetch a remote resource.
+// #[error(transparent)]
+// HttpError(#[from] reqwest::Error),
