+++ src/cache/file_downloader/impl_object_store/aws.rs
@@ -10,7 +10,7 @@ use object_store::{
 use sha2::Digest;
 use std::sync::Arc;
 
-use crate::{Error, FileDownloader, FileMetadata};
+use super::super::{Error, FileDownloader, FileMetadata};
 
 use super::{CLIENT_CONFIG_KEYS, object_store_config_key_to_string};
 
@@ -139,7 +139,7 @@ impl FileDownloader for AwsS3Downloader {
             Ok(meta) => Ok(FileMetadata {
                 size: meta.size,
                 location: meta.location.to_string(),
-                last_modified: chrono::DateTime::from(meta.last_modified),
+                last_modified: meta.last_modified,
                 e_tag: meta.e_tag,
             }),
             Err(object_store::Error::NotFound { .. }) => {
