+++ src/cache/mod.rs
@@ -1,9 +1,9 @@
-mod cache;
+mod cache_struct;
 mod error;
 mod file_downloader;
 mod meta;
 mod utils;
 
-pub use cache::*;
+pub use cache_struct::*;
 pub use error::Error;
 pub use file_downloader::*;
